+++ base-compat/CHANGES.markdown
@@ -14,6 +14,7 @@
  - Backport `inits1` and `tails1` to `Data.List.NonEmpty.Compat`.
  - Backport `minusNaturalMaybe` to `Numeric.Natural.Compat`.
  - Backport `applyWhen` to `Data.Function.Compat`.
+ - Backport `mapAccumM` and `forAccumM` to `Data.Traversable.Compat`.
 
 ## Changes in 0.12.2 [2022.08.11]
  - Sync with `base-4.17`/GHC 9.4